refactor: consolidate ChartTitle and BySource components
