+++ web/src/features/charts/NetExchangeChart.tsx
@@ -1,4 +1,5 @@
 import { useAtom } from 'jotai';
+import { useTranslation } from 'react-i18next';
 import { TimeAverages } from 'utils/constants';
 import { formatCo2 } from 'utils/formatting';
 import { displayByEmissionsAtom, productionConsumptionAtom } from 'utils/state/atoms';
@@ -19,6 +20,7 @@ function NetExchangeChart({ datetimes, timeAverage }: NetExchangeChartProps) {
   const { data, isLoading, isError } = useNetExchangeChartData();
   const [productionConsumption] = useAtom(productionConsumptionAtom);
   const [displayByEmissions] = useAtom(displayByEmissionsAtom);
+  const { t } = useTranslation();
   if (productionConsumption === 'production') {
     return null;
   }
@@ -42,7 +44,10 @@ function NetExchangeChart({ datetimes, timeAverage }: NetExchangeChartProps) {
 
   return (
     <RoundedCard className="pb-2">
-      <ChartTitle translationKey="country-history.netExchange" unit={valueAxisLabel} />
+      <ChartTitle
+        titleText={t(`country-history.netExchange.${timeAverage}`)}
+        unit={valueAxisLabel}
+      />
       <div className="relative">
         <AreaGraph
           testId="history-exchange-graph"
